+++ leetcode/src/02283_check_if_number_has_equal_digit_count_and_digit_value/solution.rs
@@ -16,7 +16,7 @@ impl Solution {
             });
         for (i, &num) in nums.iter().enumerate() {
             let cnt = *freq.get(&i).unwrap_or(&0);
-            if cnt != num as usize {
+            if cnt != num {
                 return false;
             }
         }
